test(web): co-located per-component tests — split grouped suites, add missing coverage (WidgetTypeCell, popover list, degraded channel)
